@@ -8,31 +8,31 @@ apply plugin: 'kotlin-kapt'
 apply plugin: 'io.fabric'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.3'
     flavorDimensions "default"
 
     defaultConfig {
-        applicationId "com.insperity.ontrak"
-        minSdkVersion 17
-        targetSdkVersion 28
+        applicationId "com.insperity.escmobile"
+        minSdkVersion 20
+        targetSdkVersion 29
         multiDexEnabled true
         versionCode 6
         versionName "1.2.1"
     }
 
     signingConfigs {
-        release {
-            storeFile file(ONTRAK_RELEASE_STORE_FILE)
-            storePassword ONTRAK_RELEASE_STORE_PASSWORD
-            keyAlias ONTRAK_RELEASE_KEY_ALIAS
-            keyPassword ONTRAK_RELEASE_KEY_PASSWORD
-        }
         debug {
-            keyAlias DEBUG_KEY_ALIAS
-            keyPassword DEBUG_KEY_PASSWORD
-            storeFile file(DEBUG_STORE_FILE)
-            storePassword DEBUG_STORE_PASSWORD
+            storeFile file('debug.keystore')
+            storePassword 'android'
+            keyAlias 'androiddebugkey'
+            keyPassword 'android'
+        }
+        release {
+            storeFile rootProject.file("app/release.keystore")
+            storePassword System.getProperty("APPCENTER_KEYSTORE_PASSWORD")
+            keyAlias System.getProperty("APPCENTER_KEY_ALIAS")
+            keyPassword System.getProperty("APPCENTER_KEY_PASSWORD")
         }
     }
 
