@@ -2,8 +2,8 @@ package com.insperity.escmobile
 
 import android.content.Context
 import android.os.Build
-import android.support.multidex.MultiDex
-import android.support.multidex.MultiDexApplication
+import androidx.multidex.MultiDex
+import androidx.multidex.MultiDexApplication
 import android.webkit.CookieSyncManager
 import com.crashlytics.android.Crashlytics
 import com.facebook.stetho.Stetho
