@@ -2,7 +2,7 @@ package com.insperity.escmobile.util
 
 import android.content.Context
 import android.content.res.Resources
-import android.support.v4.content.res.ResourcesCompat
+import androidx.core.content.res.ResourcesCompat
 import android.text.TextPaint
 import android.util.DisplayMetrics
 import com.insperity.escmobile.BuildConfig
