@@ -1,10 +1,10 @@
 package com.insperity.escmobile.view.adapter
 
 import android.content.Context
-import android.support.transition.TransitionManager
-import android.support.v4.content.ContextCompat
-import android.support.v7.widget.DefaultItemAnimator
-import android.support.v7.widget.RecyclerView
+import androidx.transition.TransitionManager
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.DefaultItemAnimator
+import androidx.recyclerview.widget.RecyclerView
 import android.text.Html
 import android.view.LayoutInflater
 import android.view.View
@@ -129,7 +129,8 @@ class RecentPunchesAdapter(private val context: Context?) : RecyclerView.Adapter
         if (expandedPosition == position) {
             expandedPosition = -1
             expandedId = -1
-            recyclerView.itemAnimator = DefaultItemAnimator()
+            recyclerView.itemAnimator =
+                DefaultItemAnimator()
             notifyItemChanged(position)
             return
         }
