@@ -2,7 +2,7 @@ package com.insperity.escmobile.view.common;
 
 import android.content.Context;
 import android.graphics.Typeface;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.util.TypedValue;
 
 import com.insperity.escmobile.R;
