@@ -2,7 +2,7 @@ package com.insperity.escmobile.view.fragment
 
 
 import android.os.Bundle
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.View.*
