@@ -3,8 +3,8 @@ package com.insperity.escmobile.view.fragment
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
-import android.support.design.widget.TabLayout
-import android.support.v4.app.FragmentStatePagerAdapter
+import com.google.android.material.tabs.TabLayout
+import androidx.fragment.app.FragmentStatePagerAdapter
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
