@@ -4,9 +4,10 @@ package android.support.design.internal;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.util.AttributeSet;
+import android.widget.LinearLayout;
 
 @SuppressLint("RestrictedApi")
-public class SnackbarContentLayout extends android.support.design.widget.SnackbarContentLayout {
+public class SnackbarContentLayout extends LinearLayout {
     public SnackbarContentLayout(Context context) {
         super(context);
     }
