@@ -1,6 +1,7 @@
 package com.insperity.escmobile.util
 
 import okhttp3.MediaType
+import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.RequestBody
 import okhttp3.ResponseBody
 import retrofit2.Converter
@@ -14,7 +15,7 @@ import java.lang.reflect.Type
 class ToStringConverterFactory : Converter.Factory() {
 
     companion object {
-        internal val MEDIA_TYPE = MediaType.parse("text/plain")
+        internal val MEDIA_TYPE = "text/plain".toMediaTypeOrNull()
 
         fun create() = ToStringConverterFactory()
     }
