@@ -4,7 +4,7 @@ package com.insperity.escmobile.view.fragment
 import android.animation.LayoutTransition
 import android.animation.TimeInterpolator
 import android.os.Bundle
-import android.support.design.widget.TabLayout
+import com.google.android.material.tabs.TabLayout
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
